Updated facebook.js to change id to fb.id
User.findOne({'id': profile.id}, ...) always return no user as a result every time the user login a new record for the user is added to the users's collection of mongodb database. The lookup needs to be based on 'fb.id' in order for mongodb to return a valid user (if it exist).